A couple of small fixes, enabled predicate support in windows slpd to match the other platform defaults, made the knownDA cleanup happen similarly to the property cleanups in the library.

+++ openslp/slpd/slpd_predicate.c
@@ -1490,7 +1490,7 @@ int SLPDPredicateTest(int version, size_t attrlistlen,
  *
  * @return Zero on success, or a non-zero value on failure.
  */
-int SLPDFilterAttributes(int attrlistlen, const char * attrlist,
+int SLPDFilterAttributes(size_t attrlistlen, const char * attrlist,
       size_t taglistlen, const char * taglist, size_t * resultlen, 
       char ** result)
 {
